avoid using Tempfile since it will automatically be deleted

+++ routes/wiki.rb
@@ -130,7 +130,7 @@ class MainApp < Sinatra::Base
       list = chunks[page-1].map{|x|
         x.select_attr(:id,:orig_name,:description,:size).merge({
           date:x.created_at.to_date.strftime('%Y-%m-%d'),
-          deletable: @user.admin || x.owner_id == @user.id
+          editable: @user.admin || x.owner_id == @user.id
         })
       }
       {item_id:params[:id].to_i,list: list, pages: pages, cur_page: page}
@@ -170,23 +170,48 @@ class MainApp < Sinatra::Base
     send_file(File.join(attached_base,attached.path),filename:attached.orig_name)
   end
   post '/wiki/attached/:id', private:true do
+    item = WikiItem.get(params[:id].to_i)
+    attached =  if params[:attached_id] then
+                  item.attacheds.get(params[:attached_id])
+                end
     attached_base = File.join(G_STORAGE_DIR,"attached")
-    tempfile = params[:file][:tempfile]
-    filename = params[:file][:filename]
-    date = Date.today
-    target_dir = File.join(attached_base,date.year.to_s,date.month.to_s)
-    FileUtils.mkdir_p(target_dir)
-    target_file = Tempfile.new(["attached-",".dat"],target_dir)
+    file =  if params[:file] then
+              params[:file]
+            else
+              {
+                tempfile: nil,
+                filename: params[:orig_name]
+              }
+            end
+    tempfile = file[:tempfile]
+    filename = file[:filename]
+    target_file =
+      if attached then
+        File.join(attached_base,attached.path)
+      else
+        date = Date.today
+        target_dir = File.join(attached_base,date.year.to_s,date.month.to_s)
+        FileUtils.mkdir_p(target_dir)
+        Kagetra::Utils.unique_file(@user,["attached-",".dat"],target_dir)
+      end
     res = begin
       WikiAttachedFile.transaction{
-        item = WikiItem.get(params[:id].to_i)
-        FileUtils.cp(tempfile.path,target_file)
-        rel_path = Pathname.new(target_file).relative_path_from(Pathname.new(attached_base).realpath)
-        item.attacheds.create(owner:@user,path:rel_path,orig_name:filename,description:params[:description],size:File.size(target_file))
+        FileUtils.cp(tempfile.path,target_file) if tempfile
+        if attached then
+          base = params.select_attr("description")
+          base[:orig_name] = filename
+          if tempfile then
+            base[:size] = File.size(tempfile)
+          end
+          attached.update(base)
+        else
+          rel_path = Pathname.new(target_file).relative_path_from(Pathname.new(attached_base))
+          item.attacheds.create(owner:@user,path:rel_path,orig_name:filename,description:params[:description],size:File.size(target_file))
+        end
       }
       {result:"OK"}
     rescue Exception=>e
-      FileUtils.rm(target_file)
+      FileUtils.rm(target_file) unless attached
       logger.warn e.message
       $stderr.puts e.message
       {_error_:"送信失敗"}
